+++ src/hint_processor/builtin_hint_processor/dict_hint_utils.rs
@@ -260,8 +260,7 @@ mod tests {
     use crate::hint_processor::hint_processor_definition::HintProcessor;
     use crate::stdlib::collections::HashMap;
     use crate::types::exec_scope::ExecutionScopes;
-    use crate::vm::vm_memory::memory::Memory;
-    use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
+
     use crate::{
         hint_processor::builtin_hint_processor::dict_manager::{DictManager, DictTracker},
         relocatable,