+++ src/hint_processor/builtin_hint_processor/field_arithmetic.rs
@@ -1,18 +1,19 @@
 use felt::Felt252;
-use num_bigint::BigUint;
+use num_bigint::{BigUint, ToBigInt};
+use num_integer::Integer;
 use num_traits::Zero;
 
-use crate::math_utils::{is_quad_residue, sqrt_prime_power};
+use super::hint_utils::insert_value_from_var_name;
+use super::secp::bigint_utils::Uint384;
+use crate::math_utils::{is_quad_residue, mul_inv, sqrt_prime_power};
 use crate::serde::deserialize_program::ApTracking;
 use crate::stdlib::{collections::HashMap, prelude::*};
+use crate::types::errors::math_errors::MathError;
 use crate::vm::errors::hint_errors::HintError;
 use crate::{
     hint_processor::hint_processor_definition::HintReference, vm::vm_core::VirtualMachine,
 };
 
-use super::hint_utils::{get_relocatable_from_var_name, insert_value_from_var_name};
-use super::secp::bigint_utils::BigInt3;
-use super::uint384::{pack, split};
 /* Implements Hint:
       %{
            from starkware.python.math_utils import is_quad_residue, sqrt
@@ -64,14 +65,9 @@ pub fn get_square_root(
     ids_data: &HashMap<String, HintReference>,
     ap_tracking: &ApTracking,
 ) -> Result<(), HintError> {
-    let sqrt_x_addr = get_relocatable_from_var_name("sqrt_x", vm, ids_data, ap_tracking)?;
-    let sqrt_gx_addr = get_relocatable_from_var_name("sqrt_gx", vm, ids_data, ap_tracking)?;
-    let generator = pack(
-        BigInt3::from_var_name("generator", vm, ids_data, ap_tracking)?,
-        128,
-    );
-    let x = pack(BigInt3::from_var_name("x", vm, ids_data, ap_tracking)?, 128);
-    let p = pack(BigInt3::from_var_name("p", vm, ids_data, ap_tracking)?, 128);
+    let generator = Uint384::from_var_name("generator", vm, ids_data, ap_tracking)?.pack();
+    let x = Uint384::from_var_name("x", vm, ids_data, ap_tracking)?.pack();
+    let p = Uint384::from_var_name("p", vm, ids_data, ap_tracking)?.pack();
     let success_x = is_quad_residue(&x, &p)?;
 
     let root_x = if success_x {
@@ -101,22 +97,72 @@ pub fn get_square_root(
         ids_data,
         ap_tracking,
     )?;
-    let split_root_x = split::<3>(&root_x, 128);
-    for (i, root_x) in split_root_x.iter().enumerate() {
-        vm.insert_value((sqrt_x_addr + i)?, Felt252::from(root_x))?;
-    }
-    let split_root_gx = split::<3>(&root_gx, 128);
-    for (i, root_gx) in split_root_gx.iter().enumerate() {
-        vm.insert_value((sqrt_gx_addr + i)?, Felt252::from(root_gx))?;
-    }
-
+    Uint384::split(&root_x).insert_from_var_name("sqrt_x", vm, ids_data, ap_tracking)?;
+    Uint384::split(&root_gx).insert_from_var_name("sqrt_gx", vm, ids_data, ap_tracking)?;
     Ok(())
+}
+
+/* Implements Hint:
+ %{
+    from starkware.python.math_utils import div_mod
+
+    def split(num: int, num_bits_shift: int, length: int):
+        a = []
+        for _ in range(length):
+            a.append( num & ((1 << num_bits_shift) - 1) )
+            num = num >> num_bits_shift
+        return tuple(a)
+
+    def pack(z, num_bits_shift: int) -> int:
+        limbs = (z.d0, z.d1, z.d2)
+        return sum(limb << (num_bits_shift * i) for i, limb in enumerate(limbs))
+
+    a = pack(ids.a, num_bits_shift = 128)
+    b = pack(ids.b, num_bits_shift = 128)
+    p = pack(ids.p, num_bits_shift = 128)
+    # For python3.8 and above the modular inverse can be computed as follows:
+    # b_inverse_mod_p = pow(b, -1, p)
+    # Instead we use the python3.7-friendly function div_mod from starkware.python.math_utils
+    b_inverse_mod_p = div_mod(1, b, p)
+
+
+    b_inverse_mod_p_split = split(b_inverse_mod_p, num_bits_shift=128, length=3)
+
+    ids.b_inverse_mod_p.d0 = b_inverse_mod_p_split[0]
+    ids.b_inverse_mod_p.d1 = b_inverse_mod_p_split[1]
+    ids.b_inverse_mod_p.d2 = b_inverse_mod_p_split[2]
+%}
+ */
+pub fn uint384_div(
+    vm: &mut VirtualMachine,
+    ids_data: &HashMap<String, HintReference>,
+    ap_tracking: &ApTracking,
+) -> Result<(), HintError> {
+    // Note: ids.a is not used here, nor is it used by following hints, so we dont need to extract it.
+    let b = Uint384::from_var_name("b", vm, ids_data, ap_tracking)?
+        .pack()
+        .to_bigint()
+        .unwrap_or_default();
+    let p = Uint384::from_var_name("p", vm, ids_data, ap_tracking)?
+        .pack()
+        .to_bigint()
+        .unwrap_or_default();
+
+    if b.is_zero() {
+        return Err(MathError::DividedByZero.into());
+    }
+    let b_inverse_mod_p = mul_inv(&b, &p)
+        .mod_floor(&p)
+        .to_biguint()
+        .unwrap_or_default();
+    let b_inverse_mod_p_split = Uint384::split(&b_inverse_mod_p);
+    b_inverse_mod_p_split.insert_from_var_name("b_inverse_mod_p", vm, ids_data, ap_tracking)
 }
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::hint_processor::builtin_hint_processor::hint_code;
-    use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
+    use crate::vm::errors::memory_errors::MemoryError;
     use crate::{
         any_box,
         hint_processor::{
@@ -125,12 +171,9 @@ mod tests {
             },
             hint_processor_definition::HintProcessor,
         },
-        types::{exec_scope::ExecutionScopes, relocatable::MaybeRelocatable},
+        types::exec_scope::ExecutionScopes,
         utils::test_utils::*,
-        vm::{
-            errors::memory_errors::MemoryError, runners::builtin_runner::RangeCheckBuiltinRunner,
-            vm_core::VirtualMachine, vm_memory::memory::Memory,
-        },
+        vm::vm_core::VirtualMachine,
     };
     use assert_matches::assert_matches;
 
@@ -267,5 +310,105 @@ mod tests {
             // success_x
             ((1, 15), 0)
         ];
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn run_uint384_div_ok() {
+        let mut vm = vm_with_range_check!();
+        //Initialize fp
+        vm.run_context.fp = 11;
+        //Create hint_data
+        let ids_data =
+            non_continuous_ids_data![("a", -11), ("b", -8), ("p", -5), ("b_inverse_mod_p", -2)];
+        //Insert ids into memory
+        vm.segments = segments![
+            //a
+            ((1, 0), 25),
+            ((1, 1), 0),
+            ((1, 2), 0),
+            //b
+            ((1, 3), 5),
+            ((1, 4), 0),
+            ((1, 5), 0),
+            //p
+            ((1, 6), 31),
+            ((1, 7), 0),
+            ((1, 8), 0)
+        ];
+        //Execute the hint
+        assert_matches!(run_hint!(vm, ids_data, hint_code::UINT384_DIV), Ok(()));
+        //Check hint memory inserts
+        check_memory![
+            vm.segments.memory,
+            // b_inverse_mod_p
+            ((1, 9), 25),
+            ((1, 10), 0),
+            ((1, 11), 0)
+        ];
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn run_uint384_div_b_is_zero() {
+        let mut vm = vm_with_range_check!();
+        //Initialize fp
+        vm.run_context.fp = 11;
+        //Create hint_data
+        let ids_data =
+            non_continuous_ids_data![("a", -11), ("b", -8), ("p", -5), ("b_inverse_mod_p", -2)];
+        //Insert ids into memory
+        vm.segments = segments![
+            //a
+            ((1, 0), 25),
+            ((1, 1), 0),
+            ((1, 2), 0),
+            //b
+            ((1, 3), 0),
+            ((1, 4), 0),
+            ((1, 5), 0),
+            //p
+            ((1, 6), 31),
+            ((1, 7), 0),
+            ((1, 8), 0)
+        ];
+        //Execute the hint
+        assert_matches!(
+            run_hint!(vm, ids_data, hint_code::UINT384_DIV),
+            Err(HintError::Math(MathError::DividedByZero))
+        );
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn run_uint384_div_inconsistent_memory() {
+        let mut vm = vm_with_range_check!();
+        //Initialize fp
+        vm.run_context.fp = 11;
+        //Create hint_data
+        let ids_data =
+            non_continuous_ids_data![("a", -11), ("b", -8), ("p", -5), ("b_inverse_mod_p", -2)];
+        //Insert ids into memory
+        vm.segments = segments![
+            //a
+            ((1, 0), 25),
+            ((1, 1), 0),
+            ((1, 2), 0),
+            //b
+            ((1, 3), 5),
+            ((1, 4), 0),
+            ((1, 5), 0),
+            //p
+            ((1, 6), 31),
+            ((1, 7), 0),
+            ((1, 8), 0),
+            //b_inverse_mod_p
+            ((1, 9), 0)
+        ];
+        //Execute the hint
+        assert_matches!(
+            run_hint!(vm, ids_data, hint_code::UINT384_DIV),
+            Err(HintError::Memory(MemoryError::InconsistentMemory(_, _, _)))
+        );
     }
 }