+++ src/hint_processor/builtin_hint_processor/math_utils.rs
@@ -585,7 +585,7 @@ fn div_prime_by_bound(bound: Felt252) -> Result<Felt252, VirtualMachineError> {
 mod tests {
     use super::*;
     use crate::stdlib::ops::Shl;
-    use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
+
     use crate::{
         any_box,
         hint_processor::{
@@ -599,10 +599,7 @@ mod tests {
         types::exec_scope::ExecutionScopes,
         types::relocatable::Relocatable,
         utils::test_utils::*,
-        vm::{
-            errors::memory_errors::MemoryError, runners::builtin_runner::RangeCheckBuiltinRunner,
-            vm_core::VirtualMachine, vm_memory::memory::Memory,
-        },
+        vm::{errors::memory_errors::MemoryError, vm_core::VirtualMachine},
     };
     use assert_matches::assert_matches;
     use felt::felt_str;